Comments and Sorts
Sorted the lists before they were sent anywhere. Added/Modified a few comments to hopefully make the program clearer. Made it so that the current user is added to the current list in add availabilites; up to this point the burden of adding the user to the event was on the interface.